requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.3 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
alpha-vantage==2.3.1
anyio==3.6.2
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==22.2.0 ; python_version >= '3.6'
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.88.0
fastapi-utils==0.2.1
frozenlist==1.3.3 ; python_version >= '3.7'
h11==0.14.0
httptools==0.5.0
idna==3.4
motor==3.1.1
multidict==6.0.4 ; python_version >= '3.7'
pydantic[email]==1.10.4
pymongo==4.3.3
python-decouple==3.6
python-dotenv==0.21.0
pyyaml==6.0
requests==2.28.1 ; python_version >= '3.7' and python_version < '4'
sniffio==1.3.0
sqlalchemy==1.4.46 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
starlette==0.22.0
typing-extensions==4.4.0
urllib3==1.26.13 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
uvicorn[standard]==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.2 ; python_version >= '3.7'
